Adding -s option and {{args}} macro.
Added the -s option, which allows skipping includes and defines output for
C wrappers.  Now you can use the script to generate XML files and such,
in addition to C wrappers.  Thanks to David Lecomber of Allinea for this
suggestion!

Also added {{args}} macro, which allows you to get the arguments of a function
in a list without parens.