Added missing find_package for python interpreter.
The add_wrapped_file function in the CMake module for wrap does not search for
python, but wants to use it. At the moment this is done by PnMPI, but with
searching in wrap, packages depending on wrap don't have to take care about
the internals of wrap and its CMake functions anymore.